feat(discovery): move Discovery mode inside Explore
Discovery no longer competes with Workbench as a standalone Tools
entry (May audit: questionable engagement as a primary surface):

- DiscoverySurface: the swipe/trending/leaderboard/team content
  extracted from DiscoveryPage without the page shell
- ExplorePage gains a Cheatsheets | Discovery surface switcher driven
  by ?tab=discovery, so Discovery inherits the explore feature flag
- DiscoveryPage is now a redirect shim to /explore?tab=discovery so
  old links and shortcuts keep working (routes unchanged in both apps)
- Discovery entry removed from the NavSidebar Tools section

Closes #124
